+++ src/main/java/com/webstore/service/ProductPriceService.java
@@ -1,8 +1,9 @@
 package com.webstore.service;
 
-
 import com.webstore.dto.request.ProductPriceRequestDto;
 import com.webstore.dto.response.ProductPriceResponseDto;
+
+import java.math.BigInteger;
 
 public interface ProductPriceService {
 
@@ -10,8 +11,7 @@ public interface ProductPriceService {
 
     ProductPriceResponseDto getProductPriceById(Integer id);
 
-    ProductPriceResponseDto updateProductPrice(Integer id, Long priceAmount);
+    ProductPriceResponseDto updateProductPrice(Integer id, BigInteger priceAmount);
 
     void deleteProductPrice(Integer id);
-
 }
